test(icoder): split test_app.py and fix _timed_fetch docstring
Move BranchInfoBar integration tests to test_app_branch_info.py to
keep both files under the 750-line CI cap. Share the make_icoder_app
factory via a new tests/icoder/ui/conftest.py. Add Args/Returns
sections to _timed_fetch to satisfy ruff DOC201.
